packages/commit-gen/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "commit_gen" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../coverage/commit-gen/html'
 --cov-report xml:'../coverage/commit-gen/coverage.xml'
 --html='../reports/commit-gen/unittests/html/index.html'
 --junitxml='../reports/commit-gen/unittests/junit.xml'
"""

[project]
name = "commit-gen"
version = "1.0.0"
description = "Generate commit messages from staged git changes using LLMs"
requires-python = ">=3.9,<4"
readme = 'README.md'
dependencies = [
    "gitpython>=3.1.44",
    "questionary>=1.10.0",
    "llm-connection>=1.0.0",
    "toolkit-utils>=1.0.0",
    "tqdm>=4.65.0",
    "tomli>=2.0.1"
]

[project.scripts]
cgen = "commit_gen.main:main"

[tool.hatch.build.targets.wheel]
packages = ["commit_gen"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
  "autopep8>=2.3.1",
  "ruff>=0.8.2",
  "pytest>=8.3.4",
  "pytest-sugar>=1.0.0",
  "pytest-cov>=6.0.0",
  "pytest-html>=4.1.1",
  "black>=23.3.0",
  "flake8>=5.0.4",
  "types-tqdm>=4.65.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = [
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "dist",
]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]
ignore = []

fixable = ["ALL"]
unfixable = []

[tool.black]
line-length = 80
preview = true
enable-unstable-feature = ["string_processing"]
